Refactor Contact page components
Refactor src/pages/Contact.tsx into smaller, reusable components. Remove unused imports and files.
