+++ src/components/contact/contactData.ts
@@ -1,0 +1,32 @@
+export const contactInfo = [
+  {
+    title: 'Email',
+    value: 'contacto.itopia@gmail.com',
+    icon: '📧'
+  },
+  {
+    title: 'Teléfono',
+    value: '+1 (555) 123-4567',
+    icon: '📞'
+  },
+  {
+    title: 'Ubicación',
+    value: 'Montevideo, Uruguay',
+    icon: '📍'
+  },
+  {
+    title: 'Horario',
+    value: 'Lun - Vie: 9:00 AM - 6:00 PM',
+    icon: '🕒'
+  }
+];
+
+export const serviceOptions = [
+  { value: '', label: 'Selecciona un servicio' },
+  { value: 'soporte', label: 'Soporte IT' },
+  { value: 'consultoria', label: 'Consultoría' },
+  { value: 'asesoria', label: 'Asesoría Personalizada' },
+  { value: 'modernizacion', label: 'Modernización IT' },
+  { value: 'ia', label: 'Herramientas de IA' },
+  { value: 'procesos', label: 'Mejora de Procesos' }
+];
